trying to fix differences to Frederik's work

+++ src/convert_era_5_to_seperate_basins.py
@@ -138,7 +138,7 @@ def parse_single_basin_precipitation(station_id, basin_data, discharge_file_name
     datetimes = np.concatenate(list_of_dates, axis=0)
     # concatenate the precipitation data from all the years
     precip = np.concatenate(list_of_total_precipitations, axis=0)
-    percip_mean_lat_lon = np.mean(precip, axis=(1, 2))
+    percip_mean_lat_lon = np.mean(precip, axis=(1, 2), dtype=np.float64)
     df_percip_times = pd.DataFrame(data=datetimes, index=datetimes)
     datetimes = df_percip_times.index.to_pydatetime()
     ls = [[percip_mean_lat_lon[i]] for i in range(0, len(datetimes))]
